chore: split oversized ship files for max-lines
Amp-Thread-ID: https://ampcode.com/threads/T-01a089b0-8f79-71f4-9101-bc3c5a4c95ce
